mvvm refactoring
