+++ app/src/main/java/com/example/complamap/viewmodel/SignInViewModel.kt
@@ -1,7 +1,8 @@
-package com.example.complamap
+package com.example.complamap.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
+import com.example.complamap.model.AppRepository
 
 class SignInViewModel(application: Application) : AndroidViewModel(application) {
     private var appRepository: AppRepository = AppRepository(application)
