+++ app/src/main/java/com/example/complamap/views/activities/ComplaintActivity.kt
@@ -1,4 +1,4 @@
-package com.example.complamap.activities
+package com.example.complamap.views.activities
 
 import android.os.Bundle
 import android.util.Log
