+++ app/src/main/java/com/example/complamap/views/fragments/NoAuthFragment.kt
@@ -1,4 +1,4 @@
-package com.example.complamap.fragments
+package com.example.complamap.views.fragments
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.complamap.activities.SignInActivity
+import com.example.complamap.views.activities.SignInActivity
 import com.example.complamap.databinding.FragmentNoAuthBinding
 
 class NoAuthFragment : Fragment() {
