+++ app/src/main/res/layout/fragment_mock.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context=".activities.MainActivity"
+    tools:context=".views.activities.MainActivity"
     >
 
     <TextView
